+++ client/src/components/home/ServiceInfo.tsx
@@ -63,6 +63,7 @@ const Container = styled.div<ContainerProps>`
   width: 100%;
   height: 100%;
   background-color: ${({ backgroundColor }) => backgroundColor || '#fff'};
+  margin-top: 36px;
 `;
 
 interface ImageContainerProps {
