+++ client/src/components/politicianList/LifeGraph.tsx
