+++ client/src/pages/PoliticianList.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Helmet } from 'react-helmet-async';
 import Header from '@/components/base/Header';
+import PoliticianBoard from '@/components/politicianList/PoliticianBoard';
 
 const PoliticianList = () => {
   return (
@@ -9,7 +10,7 @@ const PoliticianList = () => {
         <title>정치인 목록</title>
       </Helmet>
       <Header />
-      <div>정치인 목록</div>
+      <PoliticianBoard />
     </>
   );
 };
